refactor(gateway): restructure gateway initialization and component assembly
- Introduce a factory-based assembly path for the nanobot gateway so startup is composable and easier to test.
- Refactor `create_gateway_runtime` to accept component classes instead of hard-wiring concrete implementations.
- Update `commands.py` to start the gateway through the factory, streamlining the CLI startup flow.
- Add `nanobot/factory` modules for gateway factory wiring and related components.
- Restore cron evaluator task context under the new runtime assembly.
- Align gateway tests with factory injection.
