+++ nanobot/factory/__init__.py
@@ -1,0 +1,9 @@
+"""Runtime factories (gateway, etc.)."""
+
+from nanobot.factory.gateway import (
+    GatewayComponentClasses,
+    GatewayRuntime,
+    create_gateway_runtime,
+)
+
+__all__ = ["GatewayComponentClasses", "GatewayRuntime", "create_gateway_runtime"]
